+++ menpo/transform/__init__.py
@@ -5,3 +5,4 @@ from .piecewiseaffine import PiecewiseAffine
 from .rbf import R2LogR2RBF, R2LogRRBF
 from .groupalign.procrustes import GeneralizedProcrustesAnalysis
 from .compositions import scale_about_centre, rotate_ccw_about_centre
+from .tcoords import image_coords_to_tcoords, tcoords_to_image_coords
